requirements/base.txt
Django==1.7
argparse==1.2.1
django-bootstrap-form==3.1
djangorestframework==2.4.3
djangorestframework-csv==1.3.3
oauthlib==0.6.3
requests==2.3.0
requests-oauthlib==0.4.1
rest-pandas==0.2.0
wsgiref==0.1.2
